elixush v0.0.4 Elixush.Globals.Agent

The values defined here tend to remain constant over all runs. The atoms not starting with "global_" are used in a variety of places and therefore it is easiest to keep them global. The atoms starting with "global_" may change depending on arguments to pushgp.

Most of the values and atoms in this file are those that are used by Push instructions; all others, with few exceptions, should be defined in push_argmap in pushgp.ex and should be passed to whatever functions use them as arguments.

Functions

get_globals(key)

Specs

get_globals(atom) :: any

Get a key from the globals service

start_link(name)

update_globals(key, value)

Specs

update_globals(atom, any) :: any

Update a key in the globals service