elixush v0.0.1 Elixush.Instructions.Boolean

Functions

boolean_and(state)

boolean_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

boolean_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

boolean_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

boolean_flush(state)

Returns a function that empties the stack of the given state.

boolean_fromfloat(state)

boolean_frominteger(state)

boolean_invert_first_then_and(state)

boolean_invert_second_then_and(state)

boolean_not(state)

boolean_or(state)

boolean_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

boolean_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

boolean_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

boolean_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

boolean_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

boolean_xor(state)

boolean_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

boolean_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.