elixush v0.0.4 Elixush.Instructions.Code

Functions

code_append(state)

code_atom(state)

code_car(state)

code_cdr(state)

code_cons(state)

code_container(state)

code_contains(state)

code_do(state)

code_do_star(state)

code_do_star_count(state)

code_do_star_range(state)

code_do_star_times(state)

code_extract(state)

code_fromboolean(state)

code_fromfloat(state)

code_frominteger(state)

code_if(state)

code_insert(state)

code_length(state)

code_list(state)

code_map(state)

code_member(state)

code_noop(state)

code_nth(state)

code_nthcdr(state)

code_null(state)

code_position(state)

code_quote(state)

code_size(state)

code_subst(state)

code_wrap(state)

environment_begin(state)

Creates new environment using the entire exec stack

environment_end(state)

Ends current environment

environment_new(state)

Creates new environment using the top item on the exec stack

exec_do_star_count(state)

exec_do_star_range(state)

exec_do_star_times(state)

exec_do_star_while(state)

exec_if(state)

exec_k(state)

exec_s(state)

exec_when(state)

exec_while(state)

exec_y(state)

noop_delete_prev_paren_pair(state)

noop_open_paren(state)

positions(pred, coll)

Returns a lazy sequence containing the positions at which pred is true for items in coll.