elixush v0.0.4 Elixush.Instructions.Common

Functions

boolean_dup(state)

boolean_empty(state)

boolean_eq(state)

boolean_flush(state)

boolean_pop(state)

boolean_rot(state)

boolean_shove(state)

boolean_stackdepth(state)

boolean_swap(state)

boolean_yank(state)

boolean_yankdup(state)

char_dup(state)

char_empty(state)

char_eq(state)

char_flush(state)

char_pop(state)

char_rot(state)

char_shove(state)

char_stackdepth(state)

char_swap(state)

char_yank(state)

char_yankdup(state)

code_dup(state)

code_empty(state)

code_eq(state)

code_flush(state)

code_pop(state)

code_rot(state)

code_shove(state)

code_stackdepth(state)

code_swap(state)

code_yank(state)

code_yankdup(state)

duper(type, state)

Takes a type and a state and duplicates the top item of the appropriate stack of the state.

emptyer(type, state)

Takes a type and a state and tells whether that stack is empty.

eqer(type, state)

Compares the top two items of the type stack of the given state.

exec_dup(state)

exec_empty(state)

exec_eq(state)

exec_flush(state)

exec_pop(state)

exec_rot(state)

exec_shove(state)

exec_stackdepth(state)

exec_swap(state)

exec_yank(state)

exec_yankdup(state)

float_dup(state)

float_empty(state)

float_eq(state)

float_flush(state)

float_pop(state)

float_rot(state)

float_shove(state)

float_stackdepth(state)

float_swap(state)

float_yank(state)

float_yankdup(state)

flusher(type, state)

Empties the type stack of the given state.

integer_dup(state)

integer_empty(state)

integer_eq(state)

integer_flush(state)

integer_pop(state)

integer_rot(state)

integer_shove(state)

integer_stackdepth(state)

integer_swap(state)

integer_yank(state)

integer_yankdup(state)

popper(type, state)

Takes a type and a state and pops the appropriate stack of the state.

rotter(type, state)

Takes a type and a state and rotates the top 3 items of the appropriate stack of the state.

shover(type, state)

Shoves an item deep in the specified stack, using the top integer to indicate how deep.

stackdepther(type, state)

Pushes the depth of the type stack of the given state.

string_dup(state)

string_empty(state)

string_eq(state)

string_flush(state)

string_pop(state)

string_rot(state)

string_shove(state)

string_stackdepth(state)

string_swap(state)

string_yank(state)

string_yankdup(state)

swapper(type, state)

Takes a type and a state and swaps the top 2 items of the appropriate stack of the state.

yankduper(type, state)

Yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.

yanker(type, state)

Yanks an item from deep in the specified stack, using the top integer to indicate how deep.