elixush v0.0.4 Elixush.Instructions.Float

Functions

float_add(state)

Pushes the sum of the top two items.

float_cos(state)

float_dec(state)

float_div(state)

Returns a function that pushes the product of the top two items.

float_fromboolean(state)

float_frominteger(state)

float_fromstring(state)

float_gt(state)

Returns a function that pushes the result of comparator of the top two items on the ':float' stack onto the boolean stack.

float_gte(state)

Returns a function that pushes the result of comparator of the top two items on the ':float' stack onto the boolean stack.

float_inc(state)

float_lt(state)

Returns a function that pushes the result of comparator of the top two items on the ':float' stack onto the boolean stack.

float_lte(state)

Returns a function that pushes the result of comparator of the top two items on the ':float' stack onto the boolean stack.

float_max(state)

Returns a function that pushes the maximum of the top two items.

float_min(state)

Returns a function that pushes the minimum of the top two items.

float_mod(state)

Returns a function that pushes the modulus of the top two items. Does nothing if the denominator would be zero.

float_mult(state)

Pushes the product of the top two items.

float_sin(state)

float_sub(state)

Pushes the difference of the top two items.

float_tan(state)