elixush v0.0.1 Elixush.Instructions.Integer

Functions

integer_add(state)

Returns a function that pushes the sum of the top two items.

integer_dec(state)

integer_div(state)

Returns a function that pushes the quotient of the top two items. Does nothing if the denominator would be zero.

integer_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

integer_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

integer_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

integer_flush(state)

Returns a function that empties the stack of the given state.

integer_fromboolean(state)

integer_fromfloat(state)

integer_fromstring(state)

integer_gt(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_gte(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_inc(state)

integer_lt(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_lte(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_max(state)

Returns a function that pushes the maximum of the top two items.

integer_min(state)

Returns a function that pushes the minimum of the top two items.

integer_mod(state)

Returns a function that pushes the modulus of the top two items. Does nothing if the denominator would be zero.

integer_mult(state)

Returns a function that pushes the product of the top two items.

integer_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

integer_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

integer_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

integer_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

integer_sub(state)

Returns a function that pushes the difference of the top two items.

integer_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

integer_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

integer_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.