elixush v0.0.4 Elixush.Instructions.Vectors

Functions

butlaster(type, state)

Takes a type and a state and takes the butlast of the top item on the type stack.

concater(type, state)

Takes a type and a state and concats two vectors on the type stack.

conjer(vec_type, lit_type, state)

Takes a vec_type, a lit_type, and a state and conj's an item onto the type stack.

containser(type, lit_type, state)

Takes a type and a state and tells whether the top lit_type item is in the top type vector.

emptyvectorer(type, state)

Takes a type and a state and pushes a boolean of whether the top vector is empty.

exec_do_star_vector_boolean(state)

exec_do_star_vector_float(state)

exec_do_star_vector_integer(state)

exec_do_star_vector_string(state)

firster(type, lit_type, state)

Takes a type and a state and gets the first item from the type stack.

indexofer(type, lit_type, state)

Takes a type and a state and finds the index of the top lit_type item in the top type vector.

iterateer(type, lit_type, instr, state)

Takes a type and a state and iterates over the type vector using the code on the exec stack. If the vector isn't empty, expands to: ((first vector) (top-item :exec state) (rest vector) exec_do*vector_type (top-item :exec state) rest_of_program)

laster(type, lit_type, state)

Takes a type and a state and gets the last item from the type stack.

lengther(type, state)

Takes a type and a state and takes the length of the top item on the type stack.

nther(type, lit_type, state)

Takes a type and a state and gets the nth item from the type stack.

occurrencesofer(type, lit_type, state)

Takes a type and a state and counts the occurrences of the top lit_type item in the top type vector.

pushaller(type, lit_type, state)

Takes a type and a state and pushes every item from the first vector onto the appropriate stack.

removeer(type, lit_type, state)

Takes a type and a state and removes all occurrences of the first lit_type item in the top type vector.

replaceer(type, lit_type, state)

Takes a type and a state and replaces all occurrences of the second lit_type item with the first lit_type item in the top type vector.

replacefirster(type, lit_type, state)

Takes a type and a state and replaces the first occurrence of the second lit_type item with the first lit_type item in the top type vector.

rester(type, state)

Takes a type and a state and takes the rest of the top item on the type stack.

reverser(type, state)

Takes a type and a state and takes the reverse of the top item on the type stack.

seter(type, lit_type, state)

Takes a type and a state and replaces, in the top type vector, item at index (from integer stack) with the first lit_type item.

subvecer(type, state)

Takes a type and a state and takes the subvec of the top item on the type stack.

taker(type, state)

Takes a type and a state and takes the first N items from the type stack, where N is from the integer stack.

vector_boolean_butlast(state)

vector_boolean_concat(state)

vector_boolean_conj(state)

vector_boolean_contains(state)

vector_boolean_dup(state)

vector_boolean_empty(state)

vector_boolean_emptyvector(state)

vector_boolean_eq(state)

vector_boolean_first(state)

vector_boolean_flush(state)

vector_boolean_indexof(state)

vector_boolean_last(state)

vector_boolean_length(state)

vector_boolean_nth(state)

vector_boolean_occurrencesof(state)

vector_boolean_pop(state)

vector_boolean_pushall(state)

vector_boolean_remove(state)

vector_boolean_replace(state)

vector_boolean_replacefirst(state)

vector_boolean_rest(state)

vector_boolean_reverse(state)

vector_boolean_rot(state)

vector_boolean_set(state)

vector_boolean_shove(state)

vector_boolean_stackdepth(state)

vector_boolean_subvec(state)

vector_boolean_swap(state)

vector_boolean_take(state)

vector_boolean_yank(state)

vector_boolean_yankdup(state)

vector_float_butlast(state)

vector_float_concat(state)

vector_float_conj(state)

vector_float_contains(state)

vector_float_dup(state)

vector_float_empty(state)

vector_float_emptyvector(state)

vector_float_eq(state)

vector_float_first(state)

vector_float_flush(state)

vector_float_indexof(state)

vector_float_last(state)

vector_float_length(state)

vector_float_nth(state)

vector_float_occurrencesof(state)

vector_float_pop(state)

vector_float_pushall(state)

vector_float_remove(state)

vector_float_replace(state)

vector_float_replacefirst(state)

vector_float_rest(state)

vector_float_reverse(state)

vector_float_rot(state)

vector_float_set(state)

vector_float_shove(state)

vector_float_stackdepth(state)

vector_float_subvec(state)

vector_float_swap(state)

vector_float_take(state)

vector_float_yank(state)

vector_float_yankdup(state)

vector_integer_butlast(state)

vector_integer_concat(state)

vector_integer_conj(state)

vector_integer_contains(state)

vector_integer_dup(state)

vector_integer_empty(state)

vector_integer_emptyvector(state)

vector_integer_eq(state)

vector_integer_first(state)

vector_integer_flush(state)

vector_integer_indexof(state)

vector_integer_last(state)

vector_integer_length(state)

vector_integer_nth(state)

vector_integer_occurrencesof(state)

vector_integer_pop(state)

vector_integer_pushall(state)

vector_integer_remove(state)

vector_integer_replace(state)

vector_integer_replacefirst(state)

vector_integer_rest(state)

vector_integer_reverse(state)

vector_integer_rot(state)

vector_integer_set(state)

vector_integer_shove(state)

vector_integer_stackdepth(state)

vector_integer_subvec(state)

vector_integer_swap(state)

vector_integer_take(state)

vector_integer_yank(state)

vector_integer_yankdup(state)

vector_string_butlast(state)

vector_string_concat(state)

vector_string_conj(state)

vector_string_contains(state)

vector_string_dup(state)

vector_string_empty(state)

vector_string_emptyvector(state)

vector_string_eq(state)

vector_string_first(state)

vector_string_flush(state)

vector_string_indexof(state)

vector_string_last(state)

vector_string_length(state)

vector_string_nth(state)

vector_string_occurrencesof(state)

vector_string_pop(state)

vector_string_pushall(state)

vector_string_remove(state)

vector_string_replace(state)

vector_string_replacefirst(state)

vector_string_rest(state)

vector_string_reverse(state)

vector_string_rot(state)

vector_string_set(state)

vector_string_shove(state)

vector_string_stackdepth(state)

vector_string_subvec(state)

vector_string_swap(state)

vector_string_take(state)

vector_string_yank(state)

vector_string_yankdup(state)