elixush v0.0.3 Elixush.Instructions.Common

Functions

boolean_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

boolean_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

boolean_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

boolean_flush(state)

Returns a function that empties the stack of the given state.

boolean_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

boolean_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

boolean_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

boolean_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

boolean_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

boolean_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

boolean_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.

code_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

code_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

code_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

code_flush(state)

Returns a function that empties the stack of the given state.

code_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

code_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

code_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

code_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

code_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

code_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

code_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.

exec_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

exec_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

exec_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

exec_flush(state)

Returns a function that empties the stack of the given state.

exec_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

exec_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

exec_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

exec_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

exec_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

exec_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

exec_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.

float_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

float_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

float_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

float_flush(state)

Returns a function that empties the stack of the given state.

float_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

float_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

float_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

float_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

float_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

float_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

float_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.

integer_dup(state)

Returns a function that takes a state and duplicates the top item of the appropriate stack of the state.

integer_empty(state)

Returns a function that takes a state and tells whether that stack is empty.

integer_eq(state)

Returns a function that compares the top two items of the appropriate stack of the given state.

integer_flush(state)

Returns a function that empties the stack of the given state.

integer_pop(state)

Returns a function that takes a state and pops the appropriate stack of the state.

integer_rot(state)

Returns a function that takes a state and rotates the top 3 items of the appropriate stack of the state.

integer_shove(state)

Returns a function that shoves an item deep in the specified stack, using the top integer to indicate how deep.

integer_stackdepth(state)

Returns a function that pushes the depth of the appropriate stack of the given state.

integer_swap(state)

Returns a function that takes a state and swaps the top 2 items of the appropriate stack of the state.

integer_yank(state)

Returns a function that yanks an item from deep in the specified stack, using the top integer to indicate how deep.

integer_yankdup(state)

Returns a function that yanks a copy of an item from deep in the specified stack, using the top integer to indicate how deep.