elixush v0.0.3 Elixush.Instructions.Integer

Functions

integer_add(state)

Pushes the sum of the top two items.

integer_dec(state)

integer_div(state)

Pushes the quotient of the top two items. Does nothing if the denominator would be zero.

integer_fromboolean(state)

integer_fromfloat(state)

integer_fromstring(state)

integer_gt(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_gte(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_inc(state)

integer_lt(state)

Pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_lte(state)

Returns a function that pushes the result of comparator of the top two items on the ':integer' stack onto the boolean stack.

integer_max(state)

Returns a function that pushes the maximum of the top two items.

integer_min(state)

Returns a function that pushes the minimum of the top two items.

integer_mod(state)

Pushes the modulus of the top two items. Does nothing if the denominator would be zero.

integer_mult(state)

Pushes the product of the top two items.

integer_sub(state)

Pushes the difference of the top two items.